Test: Fix mocks after split (Phase A2)\n\n- Inline vi.mock per file to match specifiers\n- Remove shared mock helpers (path mismatch)\n- Align expectations with current service outputs\n- Keep behavior identical; fix failing offline tests\n\nRefs #534
